Project restructure.
To do away with the generated code and MPLAB X IDE Folder Structure we've migrated
to VSCode. tasks.json has everything we need to build and flash the project.

VSCode has a serial monitor we can use for serial debugging. This is a much better
environment for development.

If we can get debugging working we're laughing.