Moved PWM setup to Servo Library. This is still pretty chip dependent but its a start.
There is also an issue where if you supply the device power via the onboard
usb, it will fail to start. This might be load related, but whatever the pin
is next to the VCC supply, pulling that low will get things working again.

Added a button to pull it down. After one press after supplying power its fine.