Have a working algorithm to decide on a direction.
Using an enum to dictate the mode, the ISR functions have different behaviour
depending on the mode selected. In MOVE mode, the robot will (eventually) move
in a forward direction till it reaches an obstacle. It will then stop and
enter LOOK mode where it will scan the environment (look left and right and record
the distances). Currently, it will then look in the direction with the furthest obstacle.

Currently it will stay looking in that direction for ~5s before resetting back
to MOVE mode. This is currently controlled by a global counter variable that
increments in the ISR for the RTC (so once every second or so).